Revert "add security pointer to request_sock"
This reverts commit b21137cb9155cd5d6be91add29e8e0b9758e0c18.  Turns out
there's a cleaner way to do it, and without failing when SYN cookies are
used.  Downsides?  Labels every TCP packet after the SYNs.  Oh well,
such is the price we pay for victory.  By which I mean research.  Which
is different from victory in, well, many rather significant ways.  In
fact, it's probably more closely related to failure than to victory.
But not failure due to SYN cookies.

Conflicts:

	include/linux/security.h
	security/capability.c
	security/security.c